chore: created Utils file
It holds static methods other implementations may use
